pyproject.toml
[tool.poetry]
name = "pyproject-cookiecutter"
version = "0.1.0"
description = ""
authors = ["Sudarshan V <sudarshanv@Sudarshans-MacBook-Pro-2.local>"]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^12.5.1"
PyYAML = "^6.0"
pandas = "^1.4.3"
ttp = "^0.9.1"
openpyxl = "^3.0.10"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
pylama = "^8.3.8"
yamllint = "^1.27.1"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
